==> EY/notebooks/02_modeling.ipynb <==
!pip install pycaret

#from pycaret.regression import *
import pandas as pd

train_data = pd.read_pickle('train_dataset_v1.pkl')
train_data
train_data = train_data[np.isfinite(train_data).all(1)]

exp = setup(data = train_data, target = 'UHI Index', session_id=111)
best = compare_models()
tuned_best = tune_model(best[0], choose_better = True, early_stopping = True)
final_model = finalize_model(tuned_best)
saved_model = save_model(final_model, 'final_model_v2')
predict_model(final_model)

# predict on unseen data
loaded_model = load_model('final_model_v2')
unseen_predictions = predict_model(loaded_model, data=val_data)

submit_data['UHI Index'] = unseen_predictions['prediction_label']
