==> EY/notebooks/04_gis_analysis.ipynb <==
GIS: Geographic Information Systems
Goal: predict UHI values
Method:

- calculation of Z-scores to assess the tmeperature differences between urban and rural areas 
- utilizing LST (Land Surface Temperature) data by calcalating the mean LST and determining the standard deviation.
- UHI Intensity = (LST - Rural temp) / std
- Linear Regression / Decision Tree

To train..
* normalized Indexes (NDVI)
* training images 
--> build Linear Regression / Decision Tree Classifier --> Urban Footage Images

Urban Heat Island Classification 

**Dataset**
- sentinel-2 + ndwi
- landsat 8 + ndvi

**Parameters from Satellite**

**Alternative NDVI Adjustments**
- it reveals a nonlinear relationship with LST, emphasizing the complex role of vegetation in UHI mitigation.
